#350 Add back identity mapping on NAT interface address removal
On removing NAT interface address - Identity mapping on the interface gets to a partially
removed state in VPP. So, removing and re-adding identity mapping on interface address deletion.
A good fix would be to root cause and modify VPP and ensure the identity mappings do not need
deletion and addition (addition needed for vxlan tunnel) on nat44 address removal.